added all syntax of version 3.046+correction for parentheses bug
git-svn-id: https://www.dynare.org/svn/dynare/dynare_v4@439 ac1d8469-bf42-47a9-8791-bf33cf982152

+++ parser.src/NumericalConstants.cc
@@ -35,7 +35,7 @@ int NumericalConstants::AddConstant(string iConst)
 //------------------------------------------------------------------------------
 string NumericalConstants::get(int ID)
 {
-	if (ID < mNumericalConstants.size())
+  if (ID < (int)mNumericalConstants.size())
 	{
 		return mNumericalConstants[ID];
 	}
